Update GenerateWalletSuccessView.xaml
Due to fact that it is not known why the password is being used, best approach is to educate users.
With this change they know that password (although it might be called passphrase, but let us leave it password for now) is used as a 13th word + to encrypt wallet file locally. 
As well, added text signals to users that password is equally important as the recovery words themselves.