Promoting Matt Odell and Marty Bent
Matt and Marty have been educating the listeners of the Tales from the Crypt podcast about the importance of privacy and the nuances of Wasabi. Not only do they break down the best practices of this complex software to be accessible for regular users, but they also provide compelling arguments for privacy focused Bitcoin tools in general.
They have repeatedly proven their honor, and have well earned their part in the Dojo.