[WEBSITE] change broken link
The `free and open source` block on the website linked to the 10 commandments on verifying the PGP signatures. This is not really important in the context of why wasabi is free software. Further, since we changed the 10 commandments, this sub link is broken.

Thus this changes the hyperlink to the FAQ where it is beautifully written [why Wasabi is open source](https://docs.wasabiwallet.io/FAQ/FAQ-Introduction.html#why-is-wasabi-libre-and-open-source-software) - in my opinion a much better fit for this block.