Add new types
The new version of the protocol derives the blinding nonces using an extended master key and
an index. The coordinator has to give new nonces to the clients every time they request it.
The clients have to send the nonce back to the coordinator along with the blinded outputs so,
instead of passing public keys pack and forth it is less expensive to pass the index used in
the derivation process. In this way the required bandwidth is small and the coordinator doesn't
need to store and keep track of the nonces (it can simple derive them as it's needed)

* `PublicNonceWithIndex` is for the coordinator to send the `nonce` to the client along with the
index `N`.
* `BlindedOutputWithNonceIndex` is for the clients to send the blinded output to the coordinator
along with the index used to derive the `nonce`.