Subclass InputRequest
As we see in 7d1645605a282ba80d73d05b1953fce7993de084 during the input registration phase
the clients must send not only the blinded outputs to the coordinator but also index of
the `nonce` used as blinding factor. In this way the coordinator can derive both the public
nonce (previously sent to the client) and the private key needed to sign the blinded output.

In this commit we follow the same startegy of subclassing the `InputsRequest` by creating
an abstract `InputsRequestBase` and the two new classes: `InputsRequest` that results in the
exact same shape than before and the `InputsRequest4` class that is used to send the
blinded output + nonce index.