Subclass AliceClient
The new protocol version (4) contains two main changes, one was already addressed in the `SatoshiClient`
for getting new nonces and this one here in the `AliceClient` at the imputs registration phase. In this
case we follow again the exact same strategy consisting on subclassing `AliceClient`. The old `AliceClient`
is renamed as an abstract `AliceClientBase` with two descendats: `AliceClient` (backward compatible) and
`AliceClient4`.

The new `AliceClientBase` is exactly the same as before except in the `TryRegisterCoinsAsync` method where
a new alice is created (it can be `AliceClient` or `AliceClient4).

It is important to note that the changes in the `AliceClient.cs` file look much bigger than what they really
are because the static factory methods used to create `AliceClient` instances were unified in only one and
then overloaded to create `AliceClient4` too. The removed static factory methods were used for testing only.