Fee targets in logarithm scale
This scale is the same used by bitcoin knots (and bitcoin core soon)
to generate the fee rate histogram returned by the `getmempoolinfo`
rpc call.

Clearly it is easier to 'guess' the immediate future than the far future
and that's why it makes sense to have growing gaps in the targets and
use a log scale.

If we use the exact same scale than knots then we can use the histogram
fee rate in order to improve our estimations because the histogram tells
us how much the users are paying right now while the estimated fee
rates use past data to predict the future.