Add Equation class for linear relation proofs
This is a building block for the Sigma protocol that generalizes proof of
representation (Okamoto's protocol) and discrete log equality (Chaum-Pedersen
protocol).

A single equation is roughly equivalent to knowledge of representation but is
meant to be used internally in more general statement and knowledge classes,
that contain several equations all sharing a single witness.

This protocol is also used by the `poksho` library[1] and is described in
detail in the book "A Graduate Course in Applied Cryptography" by Boneh &
Shoup[2], which in draft version 5 is described in section 19.5.3, page 748.

[1] https://github.com/signalapp/poksho/
[2] https://toc.cryptobook.us/