Fetch Fee Rates in a RPC batch request
Makes sure to perform only one rpc request for all the standard
confirmation targets.  In case the bitcoin node fails to provide fee
rates it handles the situation and do not fail, just ignore the fee rate
that is not available.

The UTs verify this behavior by simulating RPC errors.