Fix #1: support confirmation of replaced transactions
Scenario: a user broadcast a replaceable transaction.
After a while the user decides to bump the tx fee in
order to speed up the confirmation time.
At the same the user broadcast the replacement tx the
original tx is confimed.

In this case the replacement tx has to be undone and
the replaceable one is the final valid transaction.