Fix #2: recognize replaceable txs in long chain
Scenario: a user broadcast a chain of replaceable txs
(tx1, tx2, tx3, tx4) but only tx1 is signaling rbf so,
all those transactions are replaceable while tx1 is
in the mempool. Once tx1 is confirmed tx2, tx3 and tx4
are not replaceable anymore.

This commit fixes the case where even when tx1 is still
unconfirmed, only tx2 is recognized as replaceable.