Make range proof width argument constrain exactly
Previously the width was implicitly specified by the number of bit
commitments, but now that we have an explicit argument it should be used
as the actual value, not a maximum value.

This matches the credential issuer behaviour, which enforces that the
number of bit commitments provided matches the configured width.