Needed to introduce new keypathes for new Trezor T firmware
We cannot use the old static readonly DefaultAccountKeyPath in KeyManager, because the new firmware expects a longer keypath.
The first refuse action where we expected an exception got deleted because with the new firmware there is no way to cancel or decline a Display Address action only "confirm", so no exception will be thrown by user action (maybe if we plug the device out in the meantime, but that's not ideal).