Failing test for credential issuer balance
Because `HandleRequest` validates and then mutates state, when weight
credential requests are invalid the amount credentials will have already
been processed, resulting in credentials being issued and the issuer
balance changing. This means only part of the request is processed, but
server state is still modified, which should not happen.