Add MultipartyTransaction model
Adds an immutable state model for transaction construction for use in
both the backend and the clients.

The purpose of this abstraction is to ensure that if the protocol is
followed successfully the resulting transaction will be valid, standard
and minable, and that it conforms to the parameters set in the round
itself.

Currently this is limited to only P2WPKH inputs and outputs so the rules
are simple, but this could be extended to support different script types
including support for BIPs 65 & 112, legacy inputs, and arbitrary
constraints such as only allowing standard denominations.

Caveats:
- Fee rate is enforced only for inputs & outputs, not including the
  shared overhead (up to 14.5 vbytes). Ensuring the shared overhead is
  covered can only be done once coordinator fees are implemented.
- Ownership proofs are ignored in this code, but will need to be
  accounted for eventually.
- The transaction is neither shuffled nor sorted.
- No test coverage for non-standard inputs (somewhat contrived)