Fix range proof tests in Release builds
Commit 5ba9b54c0971c264aa0160c8401c15db5fab417c changed the prover side
to only verify its own proofs in debug builds, since these checks are
very expensive for range proofs.

This means that the tests fail in release builds, since an exception is
not thrown (reported by @kiminuo).

This change removes the automatic self checking altogether and makes the
check method unconditional so that it can be called explicitly in tests
or in code that uses the prover.

Allowing the prover to create invalid proofs is perhaps not as misuse
resistant as it could be, but this makes it possible for the tests to
confirm that the verifier rejects such invalid proofs without
duplicating the prover code or manually constructing invalid range
proofs in the tests, and it is more important for the security of the
protocol that the verifier rejects invalid proofs than that the prover
refuses to create them.