Remove support for multiple inputs for now
Registering multiple inputs together is an optimization that is
appropriate for when multiple inputs with the same address are used.

This optimization is premature because it isn't currently used (the
client doesn't really exist yet) but it's making the arena and round
state much more complicated.

This should let us iterate on the critical functionality faster and
support for multiple inputs can be reintroduced after we have have a
minimal viable client working.

Co-Authored-By: lontivero <lucasontivero@gmail.com>