Add ArenaController class and HandlerProxy
Creates an `ArenaController` controller in order to have real endpoints
for our web api and test the serialization in real life.

Register default json converters in order to avoid specifying them
in every model with `[JsonConverter(Type=....)]`. In this way the models
are cleaner and easier to change. It is also easier to switch to a
different serialization mechanism in the future.