Integration test with mocked rpc
This allows us to host the coordinator in the same process and use
scoped services in order to test different scenarios that are otherway
impossible or very expensive to test. For example, it is possible
to inject services in the coordinator from the test itself what
can be used to mock services like the rpc client and run an
end-to-end test in seconds.

see: https://docs.microsoft.com/en-us/aspnet/core/test/integration-tests?view=aspnetcore-5.0