Produce coverage reports in Linux & OSX Azure pipelines
This uses the native cobertura output as described in
WalletWasabi.Documentation/CodeCoverage.md.

The `PublishCodeCoverageResults` azure pipeline task is able to generate HTML
from cobertura reports, which can be viewed in the pipeline task. For the report
generation to work `PathMap` needs to be disabled when generating the coverage
outputs.

On Windows coverage data is generated and can be downloaded for use with visual
studio, but I haven't found a way of generating HTML reports easily, so this is
not included for now.