Concurrent request handling in Arena
This partially replaces the single mutex approach used in Arena with
concurrent dictionaries for some of the state and finer grained locking.

The main benefit is that range proofs are verified without holding on to
a lock, so multiple cores can be utilized by the coordinator.

Due to violation of abstraction boundaries in tests, the AlicesById and
AlicesByOutpoint dictionaries are populated from `round.Alices` in the
periodic task. This provides backwards compatibility for the tests which
inject alices into the round/arena.