Wait until successful round marked as broadcast
The Ended phase begins during the last transaction signature request, at
which point it's inappropriate to await the RPC call, which still
happens in the periodic task.

This creates a time window during which a round may be in the Ended
phase, and the transaction state is fully signed, but it hasn't been
broadcast yet and the others' signatures are not yet known, so the round
hasn't succeeded from the point of view of the client.

Once CoinJoinClient is further integrated into the wallet services the
mempool should take precedence over this.