Fix corner case for partially discharged nodes
When a node has been partially resolved and its remaining degree is 1,
if it is chosen as a small magnitude node and appears last in the list
and the large magnitude node's balance cannot cover all of the small
magnitude nodes but no reissuance is necessary, the final edge cannot be
added as that node will not be fully discharged.

This change forces a reissuance node to be created which ensures that
node's edge will discharge the remaining balance.