Reformatting, remove FIXME comment about AsParallel
It appears that to parallelize this properly we will need to imlpement
something more refined. PLINQ can be convinced to parallelize things by
reifying some of the intermediate results, but at the expense of
consuming lots of additional memory, but merging a fixed number of
generators can be done with only a small amount of stack space.

Generating in parallel is not important for the amount decomposition
directly, might still be beneficial for computing combinations of other
users inputs during a round more quickly. If it is valuable enough, a
different approach than PLINQ will likely need to be tried.