Naive use of possible decompositions
This is a straw-man version of decomposition, intended to demonstrate to
use the possible decompositions.

The expected input spend cost for newly created outputs is not taken
into account, but only the cost of their creation.

Others' inputs amounts are not considered, nor is privacy except in that
the actual decomposition is chosen randomly from the top 100
decompositions (with up to 5000 sats or 0.01% loss permitted) ordered by
their total nominal value.