Check signature size in WabiSabi coordiantor
Taproot supports key path spending and script path spending. The
spending path cannot be determined from the script public key. The size
of witness used in the script path spending is not restricted. See
https://github.com/bitcoin/bips/blob/master/bip-0341.mediawiki#script-validation-rules

The check implemented in this commit thwarts a denial-of-service attack
in which an attacker sends a witness that is longer than expected.