Fix relative path in uri
When the base url has a subpath, and the relative url has a leading "/", the subpath of the base url is ignored. Initially ran into this in #9904, and @kiminuo showed me the correct usage. Since I was only testing with the wabisabi client, it all worked fine. Now, I was testing with the full wasabi client ans the api calls in `WasabiClient` all have a leading "/", causing the same issue. This PR removes all leading "/" in HttpClient Send requests.