Allow redirects to work in tor http client
With the upcoming BTCPay release, users will be able to to run their own coordinator.  BUT, BTCpay will not be building indexes and other wallet related services, and only offer the wabisabi api.
Ideally, if a wasabi wallet user would want to use an alternative coordinator, we would have #8852, which would allow the config to split the requests. But as a fallback, BTCpay has a catch all on all wasabi related urls, and redirects the client to the zksnacks coordinator.

The standard HttpClient handles this out of the box, but your custom Tor Tcp client magic needed redirect support to be implemented.

Most of the code used is from:
https://github.com/dotnet/runtime/blob/main/src/libraries/System.Net.Http/src/System/Net/Http/SocketsHttpHandler/RedirectHandler.cs