OutputProvider which is aware of payments
Introduce a new specialized `OutputProvider` which is aware of the existence of payments and can provide the best set of transaction outputs to be registered in a coinjoin.

It also makes sure to decompose the remaining amount (the one still available after payments) in standard denominations taking care of having the less possible effect on the AmountDecomposer's denominations selection algorithm.