Move exchange rate to client (#13028)
* Create `ExchangeRateProvider2`

* Remove old exchange rate providers and controllers

* Fetch exchange rates from the client

* Make exchange rate provider configurable

* Fix warnings

* Naming error

* CR feedback

* Use Tor to connect to providers

Most of the providers are exchanges that block access to their api from Tor. That means that not all providers can be used when Tor is enabled.

* Add `User-Agent` to requests

Some providers reject the requests if they don't have an User-Agent header. Additionally, using a constant pattern for the requests could help to finger the clients so, we use "random" user agents.

* Remove some exchange rate providers

that are unreacheable using tor. Add mempool.space as default provider.

* Add random delay between queries

---------

Co-authored-by: Kiminuo <kiminuo@protonmail.com>