Make transaction broadcasting faster.
Instead of trying to broadcast a transaction with nodes one by one, this version tries with one or more in parallel. It doesn't iterate in a loop but uses a `TaskCompletion` task to get a notification when the tx is propagated.

The number of selected nodes is one thrid of the connected nodes plus one, what makes sure to have always connected nodes (it doesn't disconnect them all). Also, this allows users to broadcast transaction immediately after the wallet starts because at that moment there use to be a very few connected nodes. Before, this was almost always impossible and the backend had to be used in most of the cases.